linq cleanup

+++ Arnible.Linq/CountExtensions.cs
@@ -4,9 +4,9 @@ namespace Arnible.Linq
 {
   public static class CountExtensions
   {
-    public static uint Count<T>(in this ReadOnlySpan<T> src, FuncIn<T, bool> func)
+    public static ushort Count<T>(in this ReadOnlySpan<T> src, FuncIn<T, bool> func)
     {
-      uint count = 0;
+      ushort count = 0;
       foreach(ref readonly T item in src)
       {
         if(func(in item))
@@ -17,9 +17,9 @@ namespace Arnible.Linq
       return count;
     }
     
-    public static uint Count<T>(in this Span<T> src, FuncIn<T, bool> func)
+    public static ushort Count<T>(in this Span<T> src, FuncIn<T, bool> func)
     {
-      uint count = 0;
+      ushort count = 0;
       foreach(ref readonly T item in src)
       {
         if(func(in item))
